+++ tests/listener.rs
@@ -19,7 +19,7 @@ fn unix_seqpacket_listener() {
 			let_assert!(Ok(mut listener) = UnixSeqpacketListener::bind(&listener));
 			async move {
 				for _ in 0..2 {
-					let_assert!(Ok((mut peer, addr)) = listener.accept().await);
+					let_assert!(Ok((peer, addr)) = listener.accept().await);
 					assert!(let None = addr.as_pathname());
 					assert!(let Ok(_) = peer.send(b"Hello!").await);
 					let mut buf = [0u8; 128];
@@ -30,7 +30,7 @@ fn unix_seqpacket_listener() {
 		});
 
 		for _ in 0..2 {
-			let_assert!(Ok(mut peer) = UnixSeqpacket::connect(&listener).await);
+			let_assert!(Ok(peer) = UnixSeqpacket::connect(&listener).await);
 			let mut buf = [0u8; 128];
 			let_assert!(Ok(len) = peer.recv(&mut buf).await);
 			assert!(&buf[..len] == b"Hello!");