+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+v0.4.0
+  * Make I/O functions take `&self` instead of `&mut self`.
+  * Deprecate the `split()` API.
+
 v0.3.0:
   * Update to tokio 0.3.2.
   * Report peer credentials with own `UCred` type since tokio made the construction private.